requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==2.1.1
cryptg==0.4.0
frozenlist==1.3.3
hachoir==3.2.0
idna==3.4
multidict==6.0.4
peewee==3.15.4
Pillow==9.4.0
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.21.1
rsa==4.9
Telethon==1.26.1
yarl==1.8.2